# node-server/api/openapi.yaml
openapi: 3.0.1
info:
  title: ToDo Manager
  description: ToDo Manager API
  version: 1.0.0
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io
servers:
- url: /
tags:
- name: tasks
  description: endpoint representing the set of all tasks
- name: users
  description: endpoint representing the set of all the users
paths:
  /api/tasks/public:
    get:
      tags:
      - tasks
      summary: Retrieve the public tasks
      description: "The public tasks (i.e., the tasks that are visible for all the\
        \ users of the service) are retrieved. This operation does not require authentication.\
        \ A pagination mechanism is implemented to limit the size of messages.\n"
      operationId: getPublicTasks
      parameters:
      - name: pageNo
        in: query
        description: "The id of the requested page (if absent, the first page is returned)"
        required: false
        style: form
        explode: true
        schema:
          type: integer
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200'
        "404":
          description: Not Found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      x-swagger-router-controller: Tasks
  /api/tasks:
    post:
      tags:
      - tasks
      summary: Create a new task
      description: A new task is created by the authenticated user.
      operationId: addTask
      requestBody:
        description: Task object that needs to be created
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
                x-content-type: application/json
        "400":
          description: Bad Request
          content: {}
        "401":
          description: Unauthorized
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
      - cookieAuth: []
      x-swagger-router-controller: Tasks
  /api/tasks/assignments:
    post:
      tags:
      - tasks
      summary: Assign automatically the unassigned tasks in a balanced manner
      description: "The tasks, whose owner is the user who performed this request\
        \ and that are not assigned to any user, are automatically assigned to the\n\
        users of the service in a balanced manner.\n"
      operationId: assign
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_1'
        "400":
          description: Bad Request
          content: {}
        "401":
          description: Unauthorized
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
      - cookieAuth: []
      x-swagger-router-controller: Tasks
  /api/tasks/{taskId}:
    get:
      tags:
      - tasks
      summary: Retreve a task
      description: The task with ID taskId is retrieved. This operation can be performed
        on the task if at least one of the following three conditions is satisfied.
        1) The task is public. 2) The user who performs the operation is the task's
        owner. 3) The user who performs the operation is a task's assignee.
      operationId: getSingleTask
      parameters:
      - name: taskId
        in: path
        description: ID of the task to retrieve
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
                x-content-type: application/json
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
      - cookieAuth: []
      x-swagger-router-controller: Tasks
    put:
      tags:
      - tasks
      summary: Update a task
      description: The task with ID taskId is updated. This operation does not allow
        to mark a task as completed. This operation can be performed only by the owner.
      operationId: updateSingleTask
      parameters:
      - name: taskId
        in: path
        description: ID of the task to update
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      requestBody:
        description: The updated task object that needs to replace the old object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
        required: true
      responses:
        "204":
          description: No Content
          content: {}
        "400":
          description: Bad Request
          content: {}
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
      - cookieAuth: []
      x-swagger-router-controller: Tasks
    delete:
      tags:
      - tasks
      summary: Delete a task
      description: The task with ID taskId is deleted. This operation can only be
        performed by the owner.
      operationId: deleteTask
      parameters:
      - name: taskId
        in: path
        description: ID of the task to delete
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Successful operation - No Content
          content: {}
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
      - cookieAuth: []
      x-swagger-router-controller: Tasks
  /api/users:
    get:
      tags:
      - users
      summary: Get information about the users
      description: |
        The available information (passwords excluded) about all the
        users is retrieved.
      operationId: getUsers
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
                x-content-type: application/json
        "401":
          description: Unauthorized
          content: {}
        "404":
          description: Not Found
          content: {}
      security:
      - cookieAuth: []
      x-swagger-router-controller: Users
  /api/users/{userId}:
    get:
      tags:
      - users
      summary: Get information about a user
      description: |
        The available information (password excluded) about the user
        specified by userId is retrieved.
      operationId: getSingleUser
      parameters:
      - name: userId
        in: path
        description: ID of the user to get
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "401":
          description: Unauthorized
          content: {}
        "404":
          description: Not Found
          content: {}
      security:
      - cookieAuth: []
      x-swagger-router-controller: Users
  /api/users/login:
    post:
      tags:
      - users
      summary: Logs a user in
      description: |
        The user who wants to log in sends the user data to the
        authenticator which performs the operation.
      operationId: loginUser
      requestBody:
        description: "The data of the user who wants to perform log in. The structure\
          \ must contain \nemail and password.\n"
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        "204":
          description: No Content
          headers:
            Set-Cookie:
              style: simple
              explode: false
              schema:
                type: string
                example: token=abcde12345; Path=/; HttpOnly
        "400":
          description: Bad Request
          content: {}
        "401":
          description: Unauthorized - Wrong password
          content: {}
        "404":
          description: Not found - Wrong username
          content: {}
      x-swagger-router-controller: Users
  /api/users/logout:
    post:
      tags:
      - users
      summary: Logs a user out
      description: |
        The user who wants to log out sends the user data to the
        authenticator which performs the operation.
      operationId: logoutUser
      requestBody:
        description: "The data of the user who wants to perform log out. The structure\
          \ must contain \nthe user id.\n"
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        "204":
          description: No Content
          content: {}
        "400":
          description: Bad Request
          content: {}
        "401":
          description: Unauthorized - Wrong password
          content: {}
        "404":
          description: Not found - Wrong username
          content: {}
      x-swagger-router-controller: Users
  /api/tasks/{taskId}/assignees:
    get:
      tags:
      - tasks
      summary: Retreve the users a task has been assigned to
      description: The users to whom the task with ID taskId is assigned are retrieved.
        This operation can be performed only by the owner.
      operationId: getUsersAssigned
      parameters:
      - name: taskId
        in: path
        description: ID of the task to retrieve
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
                x-content-type: application/json
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
      - cookieAuth: []
      x-swagger-router-controller: Tasks
    post:
      tags:
      - tasks
      summary: Assign a task to a user
      description: |
        The task with ID taskId is assigned to the user specified in the request
        body. This operation can only be performed by the owner.
      operationId: assignTaskToUser
      parameters:
      - name: taskId
        in: path
        description: ID of the task
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      requestBody:
        description: the user to whom the task is assigned
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        "204":
          description: Successful operation - No Content
          content: {}
        "400":
          description: Bad Request
          content: {}
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
      - cookieAuth: []
      x-swagger-router-controller: Tasks
  /api/tasks/{taskId}/assignees/{userId}:
    delete:
      tags:
      - tasks
      summary: Remove a user from the assigned task
      description: "The user that is removed is the user, identified by userId, that\
        \ was\nassigned to the task identified by taskId. This operation can be performed\
        \ only by the owner.\n"
      operationId: removeUser
      parameters:
      - name: taskId
        in: path
        description: ID of the assigned task
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      - name: userId
        in: path
        description: ID of the user to remove
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Successful operation - No Content
          content: {}
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
      - cookieAuth: []
      x-swagger-router-controller: Tasks
  /api/tasks/{taskId}/completion:
    put:
      tags:
      - tasks
      summary: Complete a task
      description: |
        The task with ID taskId is marked as completed. This operation can only be performed by an assignee of the task.
      operationId: completeTask
      parameters:
      - name: taskId
        in: path
        description: ID of the task
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Successful operation - No Content
          content: {}
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
      - cookieAuth: []
      x-swagger-router-controller: Tasks
  /api/users/{userId}/tasks/created:
    get:
      tags:
      - users
      summary: Retrieve the tasks created by the user with ID userId
      description: |
        This operation allows an authenticated user to retrieve the tasks that she has created. A pagination mechanism is implemented to limit the size of messages.
      operationId: getUserTasks
      parameters:
      - name: userId
        in: path
        description: ID of the user
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      - name: pageNo
        in: query
        description: "The id of the requested page (if absent, the first page is returned)"
        required: false
        style: form
        explode: true
        schema:
          type: integer
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200'
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
      - cookieAuth: []
      x-swagger-router-controller: Users
  /api/users/{userId}/tasks/assigned:
    get:
      tags:
      - users
      summary: Retrieve the tasks assigned to the user with ID userId
      description: |
        This operation allows an authenticated user to retrieve the tasks that has been assigned to her. A pagination mechanism is implemented to limit the size of messages.
      operationId: getAssignedTasks
      parameters:
      - name: userId
        in: path
        description: ID of the user
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      - name: pageNo
        in: query
        description: "The id of the requested page (if absent, the first page is returned)"
        required: false
        style: form
        explode: true
        schema:
          type: integer
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200'
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
      - cookieAuth: []
      x-swagger-router-controller: Users
components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: integer
          description: unique identifier for the task
        description:
          maxLength: 160
          type: string
          description: task description
        important:
          type: boolean
          description: "this property is set to true if the task is marked as important,\
            \ false otherwise"
          default: false
        private:
          type: boolean
          description: "this property is set to true if the task is marked as private,\
            \ false\notherwise\n"
          default: true
        project:
          type: string
          description: project in which the task is inserted
        deadline:
          type: string
          description: the due date and hour of the task
          format: date-time
        completed:
          type: boolean
          description: "this property is set to true if the task is marked as completed,\n\
            false otherwise\n"
          default: false
        owner:
          $ref: '#/components/schemas/User'
        assignees:
          type: array
          items:
            $ref: '#/components/schemas/User'
        self:
          type: string
          description: self link
          format: uri-reference
      example:
        important: false
        owner:
          password: password
          $schema: $schema
          name: name
          self: self
          id: 5
          email: ""
        private: true
        description: description
        project: project
        assignees:
        - null
        - null
        self: self
        id: 5
        completed: false
        deadline: 2000-01-23T04:56:07.000+00:00
    User:
      type: object
      properties:
        $schema:
          type: string
          description: |
            path to the JSON Schema to use for the validation of the JSON file
            describing the task
        id:
          type: integer
          description: unique identifier for the user
        name:
          type: string
          description: username of the user
        email:
          type: string
          description: email address of the user
          format: email
        password:
          maxLength: 20
          minLength: 6
          type: string
          description: password of the user
        self:
          type: string
          description: self link
          format: uri-reference
      example:
        password: password
        $schema: $schema
        name: name
        self: self
        id: 5
        email: ""
    inline_response_200:
      type: object
      properties:
        totalPages:
          type: integer
        currentPage:
          type: integer
        totalItems:
          type: integer
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        next:
          type: string
      example:
        next: next
        totalItems: 1
        totalPages: 0
        currentPage: 6
        tasks:
        - important: false
          owner:
            password: password
            $schema: $schema
            name: name
            self: self
            id: 5
            email: ""
          private: true
          description: description
          project: project
          assignees:
          - null
          - null
          self: self
          id: 5
          completed: false
          deadline: 2000-01-23T04:56:07.000+00:00
        - important: false
          owner:
            password: password
            $schema: $schema
            name: name
            self: self
            id: 5
            email: ""
          private: true
          description: description
          project: project
          assignees:
          - null
          - null
          self: self
          id: 5
          completed: false
          deadline: 2000-01-23T04:56:07.000+00:00
    inline_response_200_1:
      type: object
      properties:
        id:
          type: integer
          format: int64
      example:
        id: 0
  securitySchemes:
    cookieAuth:
      type: apiKey
      description: JWT to identify the authenticated user.
      name: connect.sid
      in: cookie